requirements.txt
aiogram==3.7.0
aiohttp==3.10.5
aiosqlite==0.20.0
python-dotenv==1.0.1
ddgs==9.0.0
